refactor compare module mixin
