+++ core/modules/compare/components/AddToCompare.ts
@@ -1,9 +1,9 @@
 import Product from '@vue-storefront/core/modules/catalog/types/Product'
-import moduleComponentMounted from '@vue-storefront/core/modules/compare/mixins/moduleComponentMounted'
+import compareMountedMixin from '@vue-storefront/core/modules/compare/mixins/compareMountedMixin'
 
 export const AddToCompare = {
   name: 'AddToCompare',
-  mixins: [moduleComponentMounted],
+  mixins: [compareMountedMixin],
   methods: {
     addToCompare (product: Product) {
       return this.$store.state['compare']
