+++ core/modules/compare/components/CompareButton.ts
@@ -1,8 +1,8 @@
-import moduleComponentMounted from '@vue-storefront/core/modules/compare/mixins/moduleComponentMounted'
+import compareMountedMixin from '@vue-storefront/core/modules/compare/mixins/compareMountedMixin'
 
 export const CompareButton = {
   name: 'CompareButton',
-  mixins: [moduleComponentMounted],
+  mixins: [compareMountedMixin],
   computed: {
     isEmpty () : boolean {
       return this.$store.getters['compare/isEmpty']
