+++ core/modules/compare/mixins/compareMountedMixin.js
